requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-filter>=24.3,<25.0
openpyxl>=3.1,<3.2
requests>=2.32,<2.33
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
